Metadata for Crossref monthly Snapshots

Crossref releases dumps of their database on the 5th of every months. These dumps sometimes change after the fact. This metadata, stored in this package, locks down the versioning of the dumps for our analyses.

md_cr_snaps_m()

update_md_cr_snaps_m()

get_md_cr_snaps_m()

Details

To lock down the reproducibility of the metadata dumps, we store their checksums in this package, instead of merely relying on their release date. (They have sometimes been updated in the past after release.)

Computing a checksum ourselves would be quite expensive Gets metadata from the AWS S3 response header.

Happily, this works without authorisation for Metadata Plus.

Current monthly snapshots:

knitr::kable(md_cr_snaps_m())

period	url	date_modified	etag	size

2018-05	https://api.crossref.org/snapshots/monthly/2018/05/all.json.tar.gz	2018-06-08 20:32:07	5fdf4c3e01a713c4391fbc193391a948-5803	48673539319

2018-06	https://api.crossref.org/snapshots/monthly/2018/06/all.json.tar.gz	2018-07-26 20:33:22	66daedd99bbac1b0406588af39654199-5913	49595814726

2018-07	https://api.crossref.org/snapshots/monthly/2018/07/all.json.tar.gz	2018-08-04 19:15:24	8b9581b551ac55e255efffed7553dc78-5937	49800224394

2018-08	https://api.crossref.org/snapshots/monthly/2018/08/all.json.tar.gz	2018-09-03 20:04:12	3250e1960d390c879cacf7ae8fe8e4e0-6006	50374663277

2018-09	https://api.crossref.org/snapshots/monthly/2018/09/all.json.tar.gz	2018-10-11 19:28:42	48dd8175cf4ef52f8452745629310e79-6153	51606812435

2018-10	https://api.crossref.org/snapshots/monthly/2018/10/all.json.tar.gz	2018-11-03 22:07:04	2de16d1ccc8fa906aa26f49d2fe752e7-6235	52296866610

2018-11	https://api.crossref.org/snapshots/monthly/2018/11/all.json.tar.gz	2018-12-03 22:02:12	123a2132ede3991a4bee12d96e4ae51c-6169	51747932929

2018-12	https://api.crossref.org/snapshots/monthly/2018/12/all.json.tar.gz	2019-01-05 01:24:48	250e3e72cc5bd4ab5dbec899b8db7580-6267	52564041493

2019-01	https://api.crossref.org/snapshots/monthly/2019/01/all.json.tar.gz	2020-02-19 19:19:18	6305e97bb00c764b1102d22c2be6d663-8495	71258741599

2019-02	https://api.crossref.org/snapshots/monthly/2019/02/all.json.tar.gz	2019-03-04 14:52:26	dd03bf567aed0e41b3736d9e9f607e85-6647	55758917898

2019-03	https://api.crossref.org/snapshots/monthly/2019/03/all.json.tar.gz	2019-04-03 14:16:02	fe2cd8dd764503640f74806e34248781-6840	57377765958

2019-04	https://api.crossref.org/snapshots/monthly/2019/04/all.json.tar.gz	2019-05-03 23:16:04	160dc0f8d12b3c91c6ed2ad0df87b4e4-7170	60145344184

2019-05	https://api.crossref.org/snapshots/monthly/2019/05/all.json.tar.gz	2019-06-04 12:16:05	1eb34559b0e80b6ad778c7a40d1dd45e-7623	63942201992

2019-06	https://api.crossref.org/snapshots/monthly/2019/06/all.json.tar.gz	2019-07-13 16:05:50	586e1ace0577ca44c422a0029960098b-7773	65197539697

2019-07	https://api.crossref.org/snapshots/monthly/2019/07/all.json.tar.gz	2019-08-04 05:46:04	31af31381ff92756beb7ab7559c33291-7858	65912914209

2019-08	https://api.crossref.org/snapshots/monthly/2019/08/all.json.tar.gz	2019-09-04 07:46:04	17782fc0071ee11ff5db519446e1edf0-7945	66646812608

2019-09	https://api.crossref.org/snapshots/monthly/2019/09/all.json.tar.gz	2019-10-04 12:42:40	196ea8897b159c0378dc5086d6fca861-8043	67464925959

2019-10	https://api.crossref.org/snapshots/monthly/2019/10/all.json.tar.gz	2019-11-13 13:42:20	5eef0f01562f7bc31d5a016fbd6042f7-5933	99530559599

2019-11	https://api.crossref.org/snapshots/monthly/2019/11/all.json.tar.gz	2019-12-05 17:16:04	3e417701f575d2a9a19582d2606231d2-8307	69683786041

2019-12	https://api.crossref.org/snapshots/monthly/2019/12/all.json.tar.gz	2020-01-06 20:39:19	66db92cdd582f8cf93b03e9b3ade167d-8424	70662719003

2020-01	https://api.crossref.org/snapshots/monthly/2020/01/all.json.tar.gz	2020-02-19 19:40:08	6305e97bb00c764b1102d22c2be6d663-8495	71258741599

2020-02	https://api.crossref.org/snapshots/monthly/2020/02/all.json.tar.gz	2020-03-05 10:21:09	cf5dbcfb9f8c4e01d7dd0d718429b642-8558	71782804006

2020-03	https://api.crossref.org/snapshots/monthly/2020/03/all.json.tar.gz	2020-04-04 08:31:09	47b39e33efd9824c7c5c68f06021ba88-8658	72624782825

2020-04	https://api.crossref.org/snapshots/monthly/2020/04/all.json.tar.gz	2020-05-11 11:52:41	104df834ffb8ff0dfd20e121a6318618-8823	74010463525

2020-05	https://api.crossref.org/snapshots/monthly/2020/05/all.json.tar.gz	2020-06-04 13:31:09	aeafd8a98336812226affb0bedc973cd-8908	74724383102

2020-06	https://api.crossref.org/snapshots/monthly/2020/06/all.json.tar.gz	2020-07-04 12:31:08	f5fbc47585eb76f74887decf4b4d36e7-9013	75604365177

2020-07	https://api.crossref.org/snapshots/monthly/2020/07/all.json.tar.gz	2020-08-04 16:31:04	1e56d469216fc77d07d103a15d9d7427-9120	76498078549

2020-08	https://api.crossref.org/snapshots/monthly/2020/08/all.json.tar.gz	2020-09-05 17:34:33	5cb71554733ac30fcb95c0b99d4c7138-9230	77423432433

2020-09	https://api.crossref.org/snapshots/monthly/2020/09/all.json.tar.gz	2020-10-27 21:16:29	1936b0825dec4b11db705f8053b7da78-9415	78973357479

2020-10	https://api.crossref.org/snapshots/monthly/2020/10/all.json.tar.gz	2020-11-05 06:01:07	ab52c1f7a2235692faa34d431751badb-9451	79275265379

2020-11	https://api.crossref.org/snapshots/monthly/2020/11/all.json.tar.gz	2020-12-05 09:31:09	7216bfd6eafcd0ff83894acece592c29-9556	80153941136

2020-12	https://api.crossref.org/snapshots/monthly/2020/12/all.json.tar.gz	2021-01-12 16:46:08	8f0977e27c695a954f843503a6120946-6367	106818152559

2021-01	https://api.crossref.org/snapshots/monthly/2021/01/all.json.tar.gz	2021-02-05 17:01:08	242406bc4cce046eaae133f4cf648978-6482	108733400373

2021-02	https://api.crossref.org/snapshots/monthly/2021/02/all.json.tar.gz	2021-03-05 18:31:09	e259608e23b932b6aed7f7c67b5256b9-6564	110120778006

2021-03	https://api.crossref.org/snapshots/monthly/2021/03/all.json.tar.gz	2021-04-05 17:31:08	9a730fb1461b0f170940845fbb93dd2f-6642	111419433519

Related Functions and Methods

Functions

update_md_cr_snaps_m: Update the metadata of monthly snapshots stored in leine

get_md_cr_snaps_m: Get metadata for currently available monthly snapshots